requirements.txt
streamlit
pandas
requests
psycopg2-binary
pypdf
reportlab
Pillow
jinja2
weasyprint
python-dotenv
boto3
